fix bug in log interface and add saver in optimfunction

+++ nrv/backend/log_interface.py
@@ -32,12 +32,12 @@ def init_reporter():
     rep_nrv._default_config['handlers']['default']['level']= 'CRITICAL'
     rep_nrv._default_config['formatters']['standard']['format']= "%(asctime)s - %(message)s"
     rep_nrv._load_defaults()
-    if MCH.do_master_only_work():
+    if MCH.do_master_only_work() and os.path.isfile("report.log"):
         os.remove("report.log")
     return rep_nrv
 
 def set_log_level(level, clear_log_file=False):
-    if clear_log_file and MCH.do_master_only_work():
+    if clear_log_file and MCH.do_master_only_work() and os.path.isfile("report.log"):
         os.remove("report.log")
     if level is not None:
         rep_nrv._default_config["loggers"][""]["level"] = level
